pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.4.1",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "leaky-bucket-py>=0.1.3",
    "aiocache>=0.12.2",
    "nltk>=3.9.1",
    "llama-index>=0.14.12",
    "llama-index-core>=0.14.12",
    "llama-index-llms-openai>=0.6.12",
    "llama-index-embeddings-openai>=0.5.1",
    "cashews[diskcache]>=7.4.4",
    "numpy>=2.3.5",
    "pytest-asyncio>=1.3.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
]

[project.scripts]
test-exercise-1 = "retrieval.exercise_1.run_tests:main"
test-exercise-2 = "retrieval.exercise_2.run_tests:main"
demo-step2-ex2 = "step2_generation.exercise_2.run_demo:main"
